forAllSystems: remove
It's not really accessible anymore because of purity concerns.
Previously, llakaLib was system-independent, but now if you need to pass
in a system first, you need to *use* forAllSystems before you can access
it, which is quite silly.

I also think flakeExposed was never the right
list to use - supporting the four main systems (the linux and darwin
ones) makes more sense.

And, the final problem - this kind of thing should really be up to the
project. I'll be using the 4 main systems on stuff like `gasp` and
`rbld`, but for my personal configuration. I'll probably just have it
with a singleton around x86_64-linux. One single function never made
much sense.

Small note: you *can* instantiate llakaLib to only have pure functions,
So, technically you could instantiate the pure version to access
forAllSystems and then instantiate the *real* version within the
forAllSystems block. But that's hell - the reason I added that feature
was just so I could say "oh yeah, there's a way to access the
functions independent of system, don't worry", and then never use it
that way.