collectDirectoryPackages: filter for derivations
We use `newScope` to allow self-reference. However, the output of
newScope adds all sorts of functions for changing the scope. As a user,
I don't want this: I just want raw derivations. I've been filtering as a
user in my projects, but this just gives me a nice attrset which I can
use for exporting.